2010 State Populations Compared to the National Average
Hover over each state to see the 2010 state population. The colors designated on the map symbolize a comparison between the state population and the national average; with red indicating a state population greater than that of the national average and pink indicating a state population lesser than that of the nation average.